Changed type from size_t to unsigned long so that header is compatible with C++

+++ planckclock.h
@@ -31,7 +31,7 @@ planck_time_t planck_time(planck_time_t*);
 
 planck_tm* planck_localtime(const planck_time_t*);
 
-size_t planck_strftime(char *s, size_t max, const char *format,
-                       const planck_tm *tm);
+unsigned long planck_strftime(char *s, unsigned long max, const char *format,
+                              const planck_tm *tm);
 
 #endif // PLANCKCLOCK_LIBRARY_H_
